+++ scratch.js
@@ -12,11 +12,9 @@ const fmt = function (iso) {
   return spacetime(iso).format('{day-short} {nice} {year}')
 }
 
-let doc = nlp('1-31')
-// let found = doc.dates({ today: [2016, 7, 13], timezone: 'Canada/Pacific' }).json()[0]
-// console.log(fmt(found.date.start))
-// console.log(fmt(found.date.end))
-
-// doc.numbers().debug().toCardinal()
-doc.numbers().debug().toOrdinal()
-console.log(doc.text())
+// let doc = nlp('july 1 2016')
+let doc = nlp('Jul 1st-31st 2016')
+// let doc = nlp('Sept 1st-30th')
+let found = doc.dates({ today: [2016, 7, 13], timezone: 'Canada/Pacific' }).json()[0]
+console.log(fmt(found.date.start))
+console.log(fmt(found.date.end))